Add support for targeting a directory
