+++ src/link/mod.rs
@@ -1,0 +1,5 @@
+pub mod link_files;
+pub mod link_options;
+
+#[cfg(test)]
+mod tests;
